+++ src/poprox_recommender/components/diversifiers/__init__.py
@@ -1,6 +1,7 @@
+from poprox_recommender.components.diversifiers.calibration import Calibrator
 from poprox_recommender.components.diversifiers.locality_calibration import LocalityCalibrator
 from poprox_recommender.components.diversifiers.mmr import MMRDiversifier
 from poprox_recommender.components.diversifiers.pfar import PFARDiversifier
 from poprox_recommender.components.diversifiers.topic_calibration import TopicCalibrator
 
-__all__ = ["MMRDiversifier", "PFARDiversifier", "TopicCalibrator", "LocalityCalibrator"]
+__all__ = ["MMRDiversifier", "PFARDiversifier", "Calibrator", "TopicCalibrator", "LocalityCalibrator"]
